riscv: add 64-bit support in scheduler
